+++ elmyra.ip.access.epo/CHANGES.rst
@@ -5,7 +5,7 @@ elmyra.ip.access.epo CHANGES
 development
 ===========
 - backpropagate current basket entries into checkbox state
-
+- display "inventor" attribute
 
 0.0.6
 =====